Reorganize any_client modules
