--- crates/cli/cli-components/src/any/mod.rs
@@ -1,4 +1,0 @@
-pub mod client_state;
-pub mod component;
-pub mod counterparty;
-pub mod queries;
