--- crates/cli/cli-components/src/counterparties/mod.rs
@@ -1,1 +1,0 @@
-pub mod tendermint;
