+++ crates/cli/cli-components/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod any;
-pub mod counterparties;
+pub mod any_client;
